Move DiagnosticProperties & DiagnosticUtils to openj9.internal.tools.attach.target
Moved classes DiagnosticProperties & DiagnosticUtils into package
openj9.internal.tools.attach.target;
Updated references accordingly.
