+++ jcl/src/java.base/share/classes/module-info.java.extra
@@ -31,7 +31,6 @@ exports com.ibm.oti.vm to java.management, jdk.attach, jdk.management, openj9.jv
 exports com.ibm.oti.util to java.management, jdk.attach, jdk.jcmd, jdk.management, openj9.sharedclasses;
 exports openj9.internal.tools.attach.target to jdk.attach, jdk.jcmd, jdk.management;
 exports openj9.management.internal to java.management, openj9.jvm;
-exports openj9.internal.tools.attach.diagnostics.base to jdk.attach, jdk.jcmd;
 exports jdk.internal.org.objectweb.asm to openj9.dtfj, openj9.dtfjview;
 // Following allows dtfj/dtfjview modules invoke module addReads & addExports programmatically via reflection APIs
 exports jdk.internal.module to openj9.dtfj, openj9.dtfjview;
