+++ jcl/src/jdk.jcmd/share/classes/openj9/tools/attach/diagnostics/tools/Jcmd.java
@@ -30,7 +30,7 @@ import java.util.List;
 import com.sun.tools.attach.VirtualMachineDescriptor;
 import com.sun.tools.attach.spi.AttachProvider;
 
-import openj9.internal.tools.attach.diagnostics.base.DiagnosticUtils;
+import openj9.internal.tools.attach.target.DiagnosticUtils;
 import openj9.tools.attach.diagnostics.attacher.AttacherDiagnosticsProvider;
 
 /**
