+++ jcl/src/jdk.jcmd/share/classes/openj9/tools/attach/diagnostics/tools/Jmap.java
@@ -27,9 +27,9 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-import openj9.internal.tools.attach.diagnostics.base.DiagnosticProperties;
-import openj9.internal.tools.attach.diagnostics.base.DiagnosticUtils;
 import openj9.internal.tools.attach.target.AttachHandler;
+import openj9.internal.tools.attach.target.DiagnosticProperties;
+import openj9.internal.tools.attach.target.DiagnosticUtils;
 import openj9.internal.tools.attach.target.IPC;
 import openj9.tools.attach.diagnostics.attacher.AttacherDiagnosticsProvider;
