+++ jcl/src/jdk.jcmd/share/classes/openj9/tools/attach/diagnostics/tools/Util.java
@@ -35,7 +35,7 @@ import java.util.Map.Entry;
 import java.util.Properties;
 import java.util.stream.Collectors;
 
-import openj9.internal.tools.attach.diagnostics.base.DiagnosticProperties;
+import openj9.internal.tools.attach.target.DiagnosticProperties;
 import openj9.internal.tools.attach.target.IPC;
 import com.sun.tools.attach.AttachNotSupportedException;
 import com.sun.tools.attach.VirtualMachine;
