+++ runtime/jcl/common/mgmtthread.c
@@ -656,7 +656,7 @@ Java_com_ibm_java_lang_management_internal_ThreadMXBeanImpl_getMultiThreadInfoIm
 }
 
 jobjectArray JNICALL
-Java_openj9_internal_tools_attach_diagnostics_base_DiagnosticUtils_dumpAllThreadsImpl(JNIEnv *env, jobject beanInstance,
+Java_openj9_internal_tools_attach_target_DiagnosticUtils_dumpAllThreadsImpl(JNIEnv *env, jobject beanInstance,
 	jboolean getLockedMonitors, jboolean getLockedSynchronizers, jint maxDepth)
 {
 	J9VMThread *currentThread = (J9VMThread *)env;
@@ -771,7 +771,7 @@ Java_com_ibm_java_lang_management_internal_ThreadMXBeanImpl_dumpAllThreadsImpl
 (JNIEnv *env, jobject beanInstance,
 	jboolean getLockedMonitors, jboolean getLockedSynchronizers, jint maxDepth)
 {
-	return Java_openj9_internal_tools_attach_diagnostics_base_DiagnosticUtils_dumpAllThreadsImpl(
+	return Java_openj9_internal_tools_attach_target_DiagnosticUtils_dumpAllThreadsImpl(
 			env, beanInstance, getLockedMonitors, getLockedSynchronizers,
 			maxDepth);
 }
