+++ runtime/jcl/exports.cmake
@@ -243,9 +243,9 @@ omr_add_exports(jclse
 	Java_com_ibm_lang_management_internal_SysinfoCpuTime_getCpuUtilizationImpl
 	Java_com_ibm_lang_management_internal_UnixExtendedOperatingSystem_getMaxFileDescriptorCountImpl
 	Java_com_ibm_lang_management_internal_UnixExtendedOperatingSystem_getOpenFileDescriptorCountImpl
-	Java_openj9_internal_tools_attach_diagnostics_base_DiagnosticUtils_getHeapClassStatisticsImpl
-	Java_openj9_internal_tools_attach_diagnostics_base_DiagnosticUtils_dumpAllThreadsImpl
-	Java_openj9_internal_tools_attach_diagnostics_base_DiagnosticUtils_triggerDumpsImpl
+	Java_openj9_internal_tools_attach_target_DiagnosticUtils_getHeapClassStatisticsImpl
+	Java_openj9_internal_tools_attach_target_DiagnosticUtils_dumpAllThreadsImpl
+	Java_openj9_internal_tools_attach_target_DiagnosticUtils_triggerDumpsImpl
 	Java_com_ibm_oti_reflect_AnnotationParser_getAnnotationsDataImpl__Ljava_lang_Class_2
 	Java_com_ibm_oti_reflect_AnnotationParser_getAnnotationsData__Ljava_lang_reflect_Constructor_2
 	Java_com_ibm_oti_reflect_AnnotationParser_getAnnotationsData__Ljava_lang_reflect_Field_2
