+++ runtime/jcl/uma/se6_vm-side_natives_exports.xml
@@ -194,8 +194,8 @@
 	<export name="Java_com_ibm_java_lang_management_internal_ThreadMXBeanImpl_findDeadlockedThreadsImpl" />
 	<export name="Java_com_ibm_java_lang_management_internal_ThreadMXBeanImpl_getMultiThreadInfoImpl" />
 	<export name="Java_com_ibm_java_lang_management_internal_ThreadMXBeanImpl_dumpAllThreadsImpl" />
-	<export name="Java_openj9_internal_tools_attach_diagnostics_base_DiagnosticUtils_getHeapClassStatisticsImpl" />
-	<export name="Java_openj9_internal_tools_attach_diagnostics_base_DiagnosticUtils_triggerDumpsImpl" />
+	<export name="Java_openj9_internal_tools_attach_target_DiagnosticUtils_getHeapClassStatisticsImpl" />
+	<export name="Java_openj9_internal_tools_attach_target_DiagnosticUtils_triggerDumpsImpl" />
 	<export name="Java_com_ibm_oti_reflect_AnnotationParser_getAnnotationsData__Ljava_lang_reflect_Field_2" />
 	<export name="Java_com_ibm_oti_reflect_AnnotationParser_getAnnotationsData__Ljava_lang_reflect_Constructor_2" />
 	<export name="Java_com_ibm_oti_reflect_AnnotationParser_getAnnotationsData__Ljava_lang_reflect_Method_2" />
@@ -317,6 +317,6 @@
 	<export name="Java_com_ibm_virtualization_management_internal_GuestOS_retrieveMemoryUsageImpl"/>
 	
 	<export name="Java_com_ibm_jvm_Stats_getStats" />
-	<export name="Java_openj9_internal_tools_attach_diagnostics_base_DiagnosticUtils_dumpAllThreadsImpl" />
+	<export name="Java_openj9_internal_tools_attach_target_DiagnosticUtils_dumpAllThreadsImpl" />
 	
 </exports>
